Ported PdfContentStreamProcessor and related classes.
Resolves: DEVSIX-318

+++ core/src/main/java/com/itextpdf/core/pdf/PdfBoolean.java
@@ -29,7 +29,7 @@ public class PdfBoolean extends PdfPrimitiveObject {
         return value;
     }
 
-    public byte getType() {
+    public int getType() {
         return Boolean;
     }
