+++ core/src/main/java/com/itextpdf/core/pdf/PdfNumber.java
@@ -31,7 +31,7 @@ public class PdfNumber extends PdfPrimitiveObject {
     }
 
     @Override
-    public byte getType() {
+    public int getType() {
         return Number;
     }
